requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
dj-database-url==2.1.0
Django==4.2.5
django-countries==7.5.1
django-filter==23.3
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.4
phonenumbers==8.13.21
Pillow==10.0.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
whitenoise==6.5.0
